Update controller and handle exceptions

+++ src/main/java/com/CstShop/ShopOnlineBackEndMain/services/basketBillServices/BasketServices.java
@@ -4,15 +4,20 @@ import com.CstShop.ShopOnlineBackEndMain.entity.basketProduct.BasketProduct;
 import com.CstShop.ShopOnlineBackEndMain.entity.products.ContentAttributes;
 import com.CstShop.ShopOnlineBackEndMain.entity.products.Products;
 import com.CstShop.ShopOnlineBackEndMain.entity.users.Users;
+import com.CstShop.ShopOnlineBackEndMain.payload.response.MessageError;
+import com.CstShop.ShopOnlineBackEndMain.payload.response.ResponseHandler;
 import com.CstShop.ShopOnlineBackEndMain.payload.response.dto.BasketProductDto;
 import com.CstShop.ShopOnlineBackEndMain.payload.response.dto.BillDto;
 import com.CstShop.ShopOnlineBackEndMain.payload.response.dto.productDtos.ProductDto;
 import com.CstShop.ShopOnlineBackEndMain.repository.basketProductRepository.BasketProductRepo;
+import com.CstShop.ShopOnlineBackEndMain.repository.productsRepository.AttributesRepo;
 import com.CstShop.ShopOnlineBackEndMain.repository.productsRepository.ContentAttributesRepo;
 import com.CstShop.ShopOnlineBackEndMain.repository.productsRepository.ProductsRepo;
 import com.CstShop.ShopOnlineBackEndMain.repository.userRepository.UsersRepo;
 import com.CstShop.ShopOnlineBackEndMain.services.productServices.TakeProductServicesImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
@@ -39,25 +44,37 @@ public class BasketServices implements BasketBillServices {
 
 	private final ContentAttributesRepo contentAttributesRepository;
 
+	private final AttributesRepo attributesRepository;
+
 
 	@Override
-	public List<BasketProductDto> addProductToBasket(Long id, Long quantity, Long contentAttributeId) {
+	public ResponseEntity<Object> addProductToBasket(Long id, Long quantity, Long contentAttributeId) {
 		Products products = productsRepository.findAllById(id);
 		if (products == null)
-			return null;
+			return ResponseHandler.generateErrorResponse(new MessageError("Insufficient quantity of products"));
+		ContentAttributes contentAttributes = contentAttributesRepository.findById(contentAttributeId).orElseThrow();
+		if (contentAttributes.getQuantity() >= quantity || quantity <= 0) {
+			return ResponseHandler.generateErrorResponse(new MessageError("Invalid product quantity"));
+		}
+		var attributeOfProduct = attributesRepository.findByProductAndContentAttributes(products, contentAttributes);
+ 		if (attributeOfProduct == null){
+			return ResponseHandler.generateErrorResponse(new MessageError("This product does not exist"));
+		}
 		BasketProduct basketProduct = new BasketProduct(quantity, products, getUser(), contentAttributeId);
 		basketProductRepository.save(basketProduct);
-		return seeAllProductFromBasket();
+		return ResponseHandler.generateResponse(ResponseHandler.MESSAGE_SUCCESS, HttpStatus.BAD_REQUEST, seeAllProductFromBasket());
 	}
 
 	@Override
-	public List<BasketProductDto> deleteProductIntoBasket(Long id) {
+	public ResponseEntity<Object> deleteProductIntoBasket(Long id) {
 		Products products = productsRepository.findAllById(id);
 		if (products == null)
-			return null;
+			return ResponseHandler.generateErrorResponse(new MessageError("Insufficient quantity of products"));
 		BasketProduct basketProduct = basketProductRepository.findAllByUserAndProduct(getUser(), products);
+		if (basketProduct == null)
+			return ResponseHandler.generateErrorResponse(new MessageError("This product does not exist"));
 		basketProductRepository.delete(basketProduct);
-		return seeAllProductFromBasket();
+		return ResponseHandler.generateResponse(ResponseHandler.MESSAGE_SUCCESS, HttpStatus.BAD_REQUEST, seeAllProductFromBasket());
 	}
 
 	@Override
@@ -84,6 +101,16 @@ public class BasketServices implements BasketBillServices {
 
 	@Override
 	public List<BillDto> seeAllBill() {
+		return null;
+	}
+
+	@Override
+	public List<BillDto> setStateAdmin(Long id, String state) {
+		return null;
+	}
+
+	@Override
+	public List<BillDto> setStateUser(Long id, String state) {
 		return null;
 	}
 }
